Tweak code for release
Fixed a number of compile warnings and removed the code that starts the daemon from GApplication (the code was not active, yet since there was a hang when checking the distributions for the compilers on x86_64 I removed the code)

+++ inst/com/src/GCOMDri.cpp
@@ -413,7 +413,7 @@ void GCOMDri::compute_dre(const GCOMObservation& obs,
             }
 
             // Compute Compton scatter angle index. Skip if it's invalid.
-            int iphibar = (event->phibar() - m_phimin) / m_phibin;
+            int iphibar = int((event->phibar() - m_phimin) / m_phibin);
             if (iphibar < 0) {
                 num_phibar_too_low++;
                 continue;
